+++ R/plot_CIF.R
@@ -1,4 +1,4 @@
-#' Plot the individual Cumulative Incidence Function (CIF) for the interest cause
+#' Plot the predictive individual Cumulative Incidence Function (CIF) for the interest cause
 #'
 #' @param DynForestPred_obj A DynForestPred object from \code{predict()} function
 #' @param id Identifiers for the selected subjects